cleaning

+++ include/BVR/physics.h
@@ -51,6 +51,6 @@ struct bvr_collision_result_s {
 void bvr_body_add_force(struct bvr_body_s* body, float x, float y, float z);
 void bvr_body_apply_motion(struct bvr_body_s* body, struct bvr_transform_s* transform);
 
-void bvr_create_collider(bvr_collider_t* collider, float* vertices, size_t count);
+void bvr_create_collider(bvr_collider_t* collider, float* vertices, uint64 count);
 void bvr_compare_colliders(bvr_collider_t* a, bvr_collider_t* b, struct bvr_collision_result_s* result);
 void bvr_destroy_collider(bvr_collider_t* collider);
